+++ src/main/java/ar/gov/sedronar/aplicacion/dto/HojaPersonalHorasDTO.java
@@ -1,17 +1,14 @@
 package ar.gov.sedronar.aplicacion.dto;
 
-import ar.gov.sedronar.aplicacion.model.Hoja;
 import ar.gov.sedronar.modulo.commonsModels.dto.ProfesionDTO;
-import ar.gov.sedronar.modulo.commonsModels.modelo.Profesion;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 /**
  * Created by TMR on 05/09/2018.
  */
 public class HojaPersonalHorasDTO {
-    private BigInteger id;
+    private Long id;
     private HojaDTO hoja;
     private ProfesionDTO profesion;
     private Integer cantidadRentadoHs;
@@ -21,11 +18,11 @@ public class HojaPersonalHorasDTO {
     private String uum;
     private java.util.Date fum;
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
