+++ src/main/java/ar/gov/sedronar/aplicacion/dto/HojasLegalesDTO.java
@@ -1,15 +1,12 @@
 package ar.gov.sedronar.aplicacion.dto;
 
-import ar.gov.sedronar.aplicacion.model.Hoja;
-
-import java.math.BigInteger;
 import java.util.Date;
 
 /**
  * Created by TMR on 05/09/2018.
  */
 public class HojasLegalesDTO {
-    private BigInteger id;
+    private Long id;
     private HojaDTO hoja;
     private String acompComisarias;
     private String acompPenales;
@@ -23,11 +20,11 @@ public class HojasLegalesDTO {
     private String uum;
     private java.util.Date fum;
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
