+++ src/main/java/ar/gov/sedronar/aplicacion/model/HojaAcompaniamientoId.java
@@ -3,7 +3,6 @@ package ar.gov.sedronar.aplicacion.model;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Objects;
 
 /**
@@ -11,22 +10,22 @@ import java.util.Objects;
  */
 @Embeddable
 public class HojaAcompaniamientoId implements Serializable {
-    private BigInteger idHoja;
+    private Long idHoja;
     private Integer idAcompaniamiento;
 
     public HojaAcompaniamientoId(){}
 
-    public HojaAcompaniamientoId(BigInteger idhoja, Integer idAcompaniamiento){
+    public HojaAcompaniamientoId(Long idhoja, Integer idAcompaniamiento){
         this.idHoja = idhoja;
         this.idAcompaniamiento = idAcompaniamiento;
     }
 
     @Column(name = "idhoja")
-    public BigInteger getIdHoja() {
+    public Long getIdHoja() {
         return idHoja;
     }
 
-    public void setIdHoja(BigInteger idHoja) {
+    public void setIdHoja(Long idHoja) {
         this.idHoja = idHoja;
     }
