+++ src/main/java/ar/gov/sedronar/aplicacion/model/HojaActividadIntervencionId.java
@@ -3,7 +3,6 @@ package ar.gov.sedronar.aplicacion.model;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Objects;
 
 /**
@@ -11,22 +10,22 @@ import java.util.Objects;
  */
 @Embeddable
 public class HojaActividadIntervencionId implements Serializable {
-    private BigInteger idHoja;
+    private Long idHoja;
     private Integer idActividad;
 
     public HojaActividadIntervencionId(){}
 
-    public HojaActividadIntervencionId(BigInteger idHoja, Integer idActividad){
+    public HojaActividadIntervencionId(Long idHoja, Integer idActividad){
         this.idHoja = idHoja;
         this.idActividad = idActividad;
     }
 
     @Column(name = "idHoja")
-    public BigInteger getIdHoja() {
+    public Long getIdHoja() {
         return idHoja;
     }
 
-    public void setIdHoja(BigInteger idHoja) {
+    public void setIdHoja(Long idHoja) {
         this.idHoja = idHoja;
     }
