+++ src/main/java/ar/gov/sedronar/aplicacion/model/HojaMensualRecursosId.java
@@ -2,9 +2,7 @@ package ar.gov.sedronar.aplicacion.model;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.EmbeddedId;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Objects;
 
 /**
@@ -13,11 +11,11 @@ import java.util.Objects;
 @Embeddable
 public class HojaMensualRecursosId implements Serializable {
     private Integer idRecursos;
-    private BigInteger idHoja;
+    private Long idHoja;
 
     public HojaMensualRecursosId(){}
 
-    public HojaMensualRecursosId(Integer idRecursos, BigInteger idhoja){
+    public HojaMensualRecursosId(Integer idRecursos, Long idhoja){
         this.idRecursos = idRecursos;
         this.idHoja = idhoja;
     }
@@ -32,11 +30,11 @@ public class HojaMensualRecursosId implements Serializable {
     }
 
     @Column(name = "idhoja")
-    public BigInteger getIdHoja() {
+    public Long getIdHoja() {
         return idHoja;
     }
 
-    public void setIdHoja(BigInteger idHoja) {
+    public void setIdHoja(Long idHoja) {
         this.idHoja = idHoja;
     }
