+++ src/main/java/ar/gov/sedronar/aplicacion/model/HojaObservaciones.java
@@ -1,7 +1,6 @@
 package ar.gov.sedronar.aplicacion.model;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 import java.util.Date;
 
 /**
@@ -10,7 +9,7 @@ import java.util.Date;
 @Entity
 @Table(name = "HOJA_OBSERVACIONES", schema ="caac")
 public class HojaObservaciones {
-    private BigInteger id;
+    private Long id;
     private Hoja hoja;
     private String alimentacionEnSede;
     private String alimentacionFueraDeSede;
@@ -31,11 +30,11 @@ public class HojaObservaciones {
 
     @Id
     @Column(name = "idhoja")
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
